+++ src/dymaptic.GeoBlazor.Core/Components/Layers/MapImageLayer.cs
@@ -93,8 +93,8 @@ public class MapImageLayer : Layer
     public MapImageLayer(string? url = null, PortalItem? portalItem = null, BlendMode? blendMode = null, 
         Dictionary<string, object>? customParameters = null, int? dpi = null, Effect? effect = null, 
         string? gdbVersion = null, MapImageFormat? imageFormat = null, int? imageMaxHeight = null,
-        int? imageMaxWidth = null, bool? imageTransparency = null, bool? legendEnabled = null, int? maxScale = null,
-        int? minScale = null, bool? persistenceEnabled = null, double? refreshInterval = null, 
+        int? imageMaxWidth = null, bool? imageTransparency = null, bool? legendEnabled = null, double? maxScale = null,
+        double? minScale = null, bool? persistenceEnabled = null, double? refreshInterval = null, 
         TimeExtent? timeExtent = null, TimeInfo? timeInfo = null, TimeInterval? timeInterval = null, 
         bool? useViewTime = null, IReadOnlyList<Sublayer>? sublayers = null, ListMode? listMode = null,
         bool? visible = null, double? opacity = null)
@@ -217,7 +217,7 @@ public class MapImageLayer : Layer
     /// </summary>
     [Parameter]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public int? MaxScale { get; set; }
+    public double? MaxScale { get; set; }
     
     /// <summary>
     ///     The minimum scale (most zoomed out) at which the layer is visible in the view. If the map is zoomed out beyond this scale, the layer will not be visible. A value of 0 means the layer does not have a minimum scale. The minScale value should always be larger than the maxScale value, and less than or equal to the service specification.
@@ -225,7 +225,7 @@ public class MapImageLayer : Layer
     /// </summary>
     [Parameter]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public int? MinScale { get; set; }
+    public double? MinScale { get; set; }
     
     /// <summary>
     ///     Refresh interval of the layer in minutes. Value of 0 indicates no refresh.
